fix(dynamic-import): resolve type alias candidates
