+++ test-parity/node-suite/module/loader/fixtures/dynamic-alias/a.ts
@@ -1,0 +1,1 @@
+export const value = "alias-a";
