+++ test-parity/node-suite/module/loader/fixtures/dynamic-alias/b.ts
@@ -1,0 +1,1 @@
+export const value = "alias-b";
